requirements.txt
pytesseract
Pillow
pdf2image
PyPDF2
python-docx
ocrmypdf
streamlit
dotenv
openai
openpyxl
pyinstaller
